Update with PR feedback
Removed the Shared projects that were breaking the build and replaced them
with folders that I include/exclude in the .csproj files.

Moved the swagger and generator files out of the top level, moved the generated
code into a directory.

Removed build properties that weren't needed and reordered things to read
a little simpler.